+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/gateways/ndelius/GetSentencesForPersonTest.kt
@@ -4,6 +4,7 @@ import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.collections.shouldBeEmpty
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
+import kotlinx.coroutines.isActive
 import org.mockito.Mockito
 import org.mockito.internal.verification.VerificationModeFactory
 import org.mockito.kotlin.verify
@@ -61,13 +62,18 @@ class GetSentencesForPersonTest(
       it("returns sentences for the matching CRN") {
         val response = nDeliusGateway.getSentencesForPerson(crn)
 
+        println("**************************************")
+        println(response.data)
+
         response.data.shouldBe(
           listOf(
             generateTestSentence(
               dateOfSentencing = LocalDate.parse("2009-07-07"),
+              isActive = false,
             ),
             generateTestSentence(
               dateOfSentencing = LocalDate.parse("2009-09-01"),
+              isActive = false,
             ),
           ),
         )
