+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/helpers/SentenceHelper.kt
@@ -5,6 +5,8 @@ import java.time.LocalDate
 
 fun generateTestSentence(
   dateOfSentencing: LocalDate? = LocalDate.parse("2020-02-03"),
+  isActive: Boolean? = true,
 ): Sentence = Sentence(
   dateOfSentencing = dateOfSentencing,
+  isActive = isActive,
 )
