+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsintegrationapi/models/ndelius/SupervisionsTest.kt
@@ -4,6 +4,9 @@ import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
 import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.Offence
 import java.time.LocalDate
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.Sentence as IntegrationApiSentence
+import uk.gov.justice.digital.hmpps.hmppsintegrationapi.models.ndelius.Sentence as NDeliusSentence
+
 
 class SupervisionsTest : DescribeSpec(
   {
@@ -94,6 +97,25 @@ class SupervisionsTest : DescribeSpec(
             ),
           )
         }
+      }
+    }
+    describe("#toSentence") {
+      it("maps one-to-one attributes to integration API sentence attributes") {
+        val supervisions = Supervisions(
+          listOf(
+            Supervision(sentence = NDeliusSentence(date = "2009-07-07", active = true)),
+            Supervision(sentence = NDeliusSentence(date = "2010-07-07", active = false)),
+          ),
+        )
+
+        val integrationApiSentences = supervisions.supervisions.map { it.toSentence() }
+
+        integrationApiSentences.shouldBe(
+          listOf(
+            IntegrationApiSentence(dateOfSentencing = LocalDate.parse("2009-07-07"), isActive = true),
+            IntegrationApiSentence(dateOfSentencing = LocalDate.parse("2010-07-07"), isActive = false),
+          ),
+        )
       }
     }
   },
